Improve separation of gui and game code
New gui functions "guiEvent*" are used to tell the gui that something changed
in the gamestate. Avoiding the game code to have to do checks for the gui and
do figure out when some gui elements have to be refreshed and how.

This patch is not complete. At least actions should not popup forms by
themselve. Buildings should have their list in the influence object, like
characters and dynasties, allowing to reference buildings by index to inform
the gui of changes. Finally it can be intersting to look at the possibility to
reserve "influence.selected" usage to the gui, since it references the player's
selection.